BAH-4821 | Refactor. Move Extensions Out Of Widgets
