+++ apps/clinical/src/extensions/__tests__/__mocks__/extensionsMocks.ts
