+++ apps/clinical/src/extensions/__tests__/utils.test.ts
