+++ apps/clinical/src/extensions/index.ts
@@ -1,0 +1,2 @@
+export type { Extension, ExtensionHandlerProps } from './models';
+export { filterByPrivileges } from './utils';
