+++ apps/clinical/src/extensions/utils.ts
