--- packages/bahmni-widgets/src/extensions/components/tabView/styles/index.module.scss
@@ -1,11 +1,0 @@
-@use "@carbon/react/scss/spacing" as *;
-@use "@carbon/react/scss/type" as *;
-@use "@carbon/react/scss/theme" as *;
-
-.message {
-  @include type-style("body-compact-01");
-  color: $support-error;
-  font-style: italic;
-  text-align: center;
-  padding: $spacing-10;
-}
