--- packages/bahmni-widgets/src/extensions/index.ts
@@ -1,16 +1,0 @@
-import '../search';
-
-export { default as Extensions } from './Extensions';
-export {
-  registerExtensionWidget,
-  getExtensionWidget,
-  clearExtensionWidget,
-  type ExtensionWidget,
-  type ExtensionWidgetProps,
-} from './registry';
-export { useExtensionConfig } from './context';
-export type {
-  Extension,
-  ExtensionParams,
-  ExtensionConfigContextType,
-} from './models';
