--- packages/bahmni-widgets/src/extensions/registry/index.ts
@@ -1,6 +1,0 @@
-export {
-  registerExtensionWidget,
-  getExtensionWidget,
-  clearExtensionWidget,
-} from './registry';
-export type { ExtensionWidget, ExtensionWidgetProps } from './models';
