--- packages/bahmni-widgets/src/extensions/registry/models.ts
@@ -1,8 +1,0 @@
-export interface ExtensionWidgetProps {
-  configUrl?: string;
-}
-
-export interface ExtensionWidget {
-  key: string;
-  component: React.ComponentType<ExtensionWidgetProps>;
-}
