--- packages/bahmni-widgets/src/extensions/schema.json
@@ -1,29 +1,0 @@
-{
-  "$schema": "http://json-schema.org/draft-07/schema#",
-  "title": "Search Extension Configuration Schema",
-  "type": "array",
-  "items": {
-    "type": "object",
-    "additionalProperties": true,
-    "required": ["id", "extensionPointId", "type", "translationKey", "extensionParams"],
-    "properties": {
-      "id": { "type": "string" },
-      "extensionPointId": { "type": "string" },
-      "type": { "type": "string" },
-      "translationKey": { "type": "string" },
-      "icon": { "type": "string" },
-      "requiredPrivileges": {
-        "type": "array",
-        "items": { "type": "string" }
-      },
-      "extensionParams": {
-        "type": "object",
-        "required": ["searchHandler", "configUrl"],
-        "properties": {
-          "searchHandler": { "type": "string" },
-          "configUrl": { "type": "string" }
-        }
-      }
-    }
-  }
-}
